chore: Add Gantt chart and PERT diagram LaTeX files
This commit adds the Gantt chart and PERT diagram LaTeX files to the `.report/misc` directory. The Gantt chart is added in the `gantt.tex` file, while the PERT diagram is added in the `pert.tex` file. These additions enhance the documentation by providing visual representations of project timelines and dependencies.