chore: Add .report directory to .gitignore
The commit adds the `.report` directory to the `.gitignore` file, preventing it from being tracked by Git. This directory is used for storing LaTeX files and other miscellaneous files related to the project's documentation. By ignoring it, we ensure that these files do not clutter the repository and are not accidentally committed.